Assassin buffs, Combat classes with hunted beware (#4179)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
Adds a new garrote specially for assassin, and gives them a lock pick
stash
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
For too long have people picked hunted to just play a combat class,
thinking that they are safe. For too long have we had to noisely stab
people as assassin and have your target tell everyone in a 50 mile
radius that you are an assassin.
This PR changes this. Basically, assassins get an unbreakable
garrote(sprites pending), obtainable via dagger stash, along with a lock
pick stash. These 2 small changes instantly make assassins a lot scarier
to face alone, but still easy to beat when you have more than one person
fighting them. This also gives assassins a way to escape confinement
when captured if there is a stash available, leading to "OH GODS WHERE
IS HE, HE COULD BE ANYWHERE" moments.
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->

## Changelog

<!-- If your PR modifies aspects of the game that can be concretely
observed by players or admins you should add a changelog. If your change
does NOT meet this description, remove this section. Please note that
maintainers freely reserve the right to remove and add tags should they
deem it appropriate. You can attempt to finagle the system all you want,
but it's best to shoot for clear communication right off the bat. -->

<!-- !! Do not add whitespace in-between the entries, do not change the
tags and do not leave the tags without any entries. -->
:cl:
add: Added Profane Razor,  a new tool for assassins
add: Added lock picks to the assassins stash
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [x] You tested this on a local server.
- [x] This code did not runtime during testing.
- [x] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->