(SOUND) Adds a new theme for the Sacresants (#4073)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request

Adds the song "TRINITY (You Cant Kill Me)" by T87-Sulfurhead to the
Sacresants.

## Why It's Good For The Game

God said I can be Blade for a day. I love Blade III: Trinity so much.

## Changelog

<!-- !! Do not add whitespace in-between the entries, do not change the
tags and do not leave the tags without any entries. -->
:cl:
add:Adds a new music track for the Sacresants.
/:cl:
## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [x] You tested this on a local server.
- [x] This code did not runtime during testing.
- [x] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->