Inquisition weapon stat changes + No more defense from enchant (#4121)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
A lot of this shit had crazy stats like 8 defense on a halberd (when 5
is supposed to be the max for a weapon), because the inquisition enchant
just gave defense levels like biscuits

-Inquisition censer enchant doesn't give defense anymore

-Katar force and pen nerf
The katar does way too much damage and easily delimbs people
Pen reduced to 30 / Force reduced to 15
defense reduced to 2

Fix Creed being really weird (estoc intent, warhammer impale that just
pierces through plate armor like butter)
-Creed now inherits from the longsword instead of the greatsword.
-Removed the random intents from creed.

-Unforgotten blade inherits from psydonian greatsword (unnused sword)

-Change mapped in creed path to new one
## Why It's Good For The Game

<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->

## Changelog

<!-- If your PR modifies aspects of the game that can be concretely
observed by players or admins you should add a changelog. If your change
does NOT meet this description, remove this section. Please note that
maintainers freely reserve the right to remove and add tags should they
deem it appropriate. You can attempt to finagle the system all you want,
but it's best to shoot for clear communication right off the bat. -->

<!-- !! Do not add whitespace in-between the entries, do not change the
tags and do not leave the tags without any entries. -->
:cl:
add: Changed creed's parent and removed some of its intents
balance: Inquisition enchant doesn't give defense anymore
balance: reduced katar force and pen a bit
map: change path of a mapped in weapon in the bog
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [x] You tested this on a local server.
- [x] This code did not runtime during testing.
- [x] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->