"Call it a reliquary for Psydon's sake" : Un-frenches the inquisition's hermes (#4290)
## About The Pull Request
Changes the text in the inquisition's reliquary to not be in French.
Also fixes the way the sentence is formed, which was nonsensical even
translated in english.

Despite my best attempts, I could not fix the UTF-8 characters showing
up as gibberish.

## Why It's Good For The Game

_"Pour l'éradication de l'hérésie, tant que Psydon endvre" ? Passe le
jeu en anglais, merde !_

## Changelog

:cl:
fix: The Oratorium reliquary menu no longer is French
/:cl:

## Vérifications avant l'ajout

- [x] Vous avez testé votre code sur un serveur local.
- [x] Ce code n'a pas runtime durant l'exécution.
- [x] Vous avez documenté tous vos changements.

Co-authored-by: Somniworld <pablo.capel4@orange.fr>