(Please Testmerge first) Gives necran a new miracle, "Locate corpse" (#4230)
## About The Pull Request

Gives Necran clergy a new T2 miracle named "locate corpse", code
originally made by drsirknight (PR #3635).
It replaces Speak with soul, which this PR axes, and allows necran
clergy to locate corpses with a 1 minute cooldown, and a cost of 50
devotion.

## Why It's Good For The Game
Playing gravekeeper is cool, but half of the deaths happen in super
vague locations like "a locale wreathed in enigmatic fog", which can
mean the basin, or the tomb of matthios, or somewhere else altogether.
Thus, most corpses are a pain in the ass to find. This new miracle will
help in this regard.

~~Speak with soul is cool as a concept, but it halts people respawning,
and a ghost who just got pulled from the great beyond often wonders if
they're able to communicate information they've seen while they were
observing as a ghost, and so on. I think it's better if ghosts who
depart for the underworld are gone forever. This is up for debate,
though, and I'll listen to your input if you disagree.~~
Nevermind, Speak With Soul stays.

## Video showcase
I was fucking around with admin tools while testing this, this is why I
am playing a wretch class with necran miracles



https://github.com/user-attachments/assets/6c62df92-8b68-4cfb-90f0-325235e0a087


## Changelog


:cl:
add: Grants Necran clergy a new T2 miracle : Locate corpse, which allows
them to find bodies, though said body may be a deadite.



/:cl:



## Pre-Merge Checklist
I'm confident it works, but I did not test leaving a corpse hanging for
30+ minutes to see if the affix would change in the list. I also did not
test corpses being in other areas, and would like to check it with a
test merge, please.


- [x] You tested this on a local server.
- [x] This code did not runtime during testing.
- [x] You documented all of your changes.

---------

Co-authored-by: Somniworld <pablo.capel4@orange.fr>
Co-authored-by: Washington <47699733+wgtjunior743@users.noreply.github.com>