Remove defunct dismember sound effect code (#4301)
`"desceration"` does not exist in `get_sfx()`, so it results in no sound
and any code referencing it is pointless. also, `dismember()` uses the
bodypart's dismember sound. also it's a typo of `"desecration"` anyway.

due to this `get_dismember_sound()` would always return a blank/invalid
sound, so it was removed as well. dismember sounds are handled elsewhere
now anyway