Caps Grenzelhoft Mercenary (THE GUNCONOMY IS IN SHAMBLES) (#4338)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request

Caps Grenzelhoftian Mercenary from 5->2. This is because they now use a
musket.

## Why It's Good For The Game

Guns are supposed to be rare, and their powder is supposed to be kept a
strict secret. Please don't have a morbillion muskets laying around,
they're meant to be powerful, and ranged weaponry is already busted
as-is. This doesn't make sense from a lore perspective, and it doesn't
feel right balancing wise.

I want more of my stupid Grenz mercenaries running around but until
someone makes a class separately for it (FOR THE LOVE OF GOD DON'T WE
HAVE SO MUCH MERCENARY BLOAT) we gotta have a limit for it.

## Changelog

<!-- !! Do not add whitespace in-between the entries, do not change the
tags and do not leave the tags without any entries. -->
:cl:
balance: Caps Grenzelhoft Mercenary to 2, due to his musket.
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [x] You tested this on a local server.
- [x] This code did not runtime during testing.
- [x] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->