Indentured in Style (Servant Outfits + Service Bell) (#4271)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->
This adds several new pieces of clothing to the game (spritework done by
Cre). These are primarily for the manor servants, but some are ambiguous
enough for other uses. The clothing items are:
- Formal Shirt
- Formal Trousers
- Trouser Shorts
- Tailcoat
- Suspenders (Belt slot)
- Maid Headband
- Maid Apron
- Maid Dress
- Servant Gown
- Cloth Belt (not a sash)

<img width="495" height="396" alt="image"
src="https://github.com/user-attachments/assets/5e4b3cf1-f1f9-434c-866d-af20f85a8ea9"
/>


In addition to this, a service bell has been added to the game. 1 is
given to the butler, but several have been added to the map in the
lord's chambers, heirs' chambers, hand's chambers, and archivist's
chambers. When rang, it acts similarly to a signal horn for the
servants. If they're within the keep (give or take), they will receive
an indicator of the sound's direction, as well as if it was up or down
if it was on a different level. This has a 1 minute cooldown for
nobility, and 3 minutes for non-nobility, though you can ring it as much
as you like without the effect.
Ring sound effect was done by fem_tanyl

## Why It's Good For The Game

<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->
i was asked to do it and so it has to be good, because why else would
they ask me to do it??? :clueless:


## Changelog

<!-- If your PR modifies aspects of the game that can be concretely
observed by players or admins you should add a changelog. If your change
does NOT meet this description, remove this section. Please note that
maintainers freely reserve the right to remove and add tags should they
deem it appropriate. You can attempt to finagle the system all you want,
but it's best to shoot for clear communication right off the bat. -->

<!-- !! Do not add whitespace in-between the entries, do not change the
tags and do not leave the tags without any entries. -->
:cl:
add: New Servant/Butler Outifts
add: Servant Bell
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [x] You tested this on a local server.
- [x] This code did not runtime during testing.
- [x] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->

---------

Co-authored-by: CheffieGithub <113442598+CheffieGithub@users.noreply.github.com>