Rework vampire insta kill when victim is out of vitae (#4405)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
This makes it so instead of your victim instantly dying, they go
unconscious for 50 ish seconds.

Closes https://github.com/Monkestation/Vanderlin/issues/3813
Closes https://github.com/Monkestation/Vanderlin/issues/3933
Closes https://github.com/Monkestation/Vanderlin/issues/4044
Closes https://github.com/Monkestation/Vanderlin/issues/4403
Closes https://github.com/Monkestation/Vanderlin/issues/4402
Closes https://github.com/Monkestation/Vanderlin/issues/4112

## Why It's Good For The Game
This really sucks, you can lag and take one extra sip anddd the guy
instantly dies.

It also made feeding with more than 1 vampire pretty much impossible
because the guy you're feeding on would instantly die as well.

## Changelog

<!-- If your PR modifies aspects of the game that can be concretely
observed by players or admins you should add a changelog. If your change
does NOT meet this description, remove this section. Please note that
maintainers freely reserve the right to remove and add tags should they
deem it appropriate. You can attempt to finagle the system all you want,
but it's best to shoot for clear communication right off the bat. -->

<!-- !! Do not add whitespace in-between the entries, do not change the
tags and do not leave the tags without any entries. -->
:cl:
fix: vampires don't instantly kill someone if they suck when victim runs
out of vitae.
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [x] You tested this on a local server.
- [x] This code did not runtime during testing.
- [x] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->

---------

Co-authored-by: Washington <47699733+wgtjunior743@users.noreply.github.com>