Makes the Locate Dead miracle less expensive. (#4365)
## About The Pull Request

Changes the devotion cost of the "find corpse" miracle from 50 to 15.
And it's cooldown from 60 seconds to 30 seconds.

## Why It's Good For The Game

Locate corpse is an essential tool to gravekeepers, but it costing 50
devotion means it takes one entire minute to muster the necessary
resources to cast it. It also has a long cooldown despite being
(purposefully) vague. Buffing it will make it less of a chore to find
bodies.

## Changelog

:cl:

balance: the Find Corpse Necran miracle now costs 15 devotion, down from
50, it also has a CD of 30 seconds, down from 60.

/:cl:
- [x] You tested this on a local server.
- [x] This code did not runtime during testing.
- [x] You documented all of your changes.

Co-authored-by: Somniworld <pablo.capel4@orange.fr>