Court Agent List Fixes/Polish (#4435)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->
This PR messes with the Frumentarii system for court agents and the
list. This should resolve an issue where the Hand would not know their
agents and their list would not populate. Additionally, changed some
aspects of the list itself.
- Made the scroll fireproof
- Added a new variable to paper, writeable. This makes it so trying to
use a writing utensil on it will not prompt you for contents. This is on
by default and disabled for frumentarii lists.
- The Frumentarii list only has space for 5 names. Because they're
fingers to a Hand. Come on.
- To add to the scroll you must be a noble. You also cannot add members
of the royal family to the fingers.
- Redid the UI for severing fingers, it now uses browser_input.
Removed Agents are not removed from memory. Instead, frumentarii are now
stored as an assoc of name = TRUE/FALSE, where the bool indicates
whether they are presently frumentarii. If they are not, they have their
own examine text.

## Why It's Good For The Game

<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->
Fixes and polish. Nicer to use, bugs fixed.

## Changelog

<!-- If your PR modifies aspects of the game that can be concretely
observed by players or admins you should add a changelog. If your change
does NOT meet this description, remove this section. Please note that
maintainers freely reserve the right to remove and add tags should they
deem it appropriate. You can attempt to finagle the system all you want,
but it's best to shoot for clear communication right off the bat. -->

<!-- !! Do not add whitespace in-between the entries, do not change the
tags and do not leave the tags without any entries. -->
:cl:
fix: Changes to how the Hand's List works, Fixes for Hand not
recognizing their agents.
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [x] You tested this on a local server.
- [x] This code did not runtime during testing.
- [x] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->

---------

Co-authored-by: CheffieGithub <113442598+CheffieGithub@users.noreply.github.com>