Gives Twitch subs Donator perks, adds give triumphs Plexora Topic (#4519)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

Twitch Integration PR, gives anyone subbed to TheDukeofOOK on Twitch
same donator perks as patrons. Keeps us in line with Monke Main

Also adds a new Plexora topic for awarding triumphs. The only current
use for this is for Ook's Twitch bot, but could easily be used with
Plexora itself.

Of course not included in the PR but already completed is updating our
DB with twitch ranks. This mirrors anyone linking their ckey on Monke
main, anyone who wishes to link their Twitch should continue to do so on
Monkestation.

Lastly, changes a lot of the "patreon" checks to "donator" checks.
## Why It's Good For The Game

<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->

Gives an alternative to Patreon for all of our donator perks, and a
"free" option to those who have Amazon Prime and haven't used their
Twitch Prime Sub.

## Changelog

<!-- If your PR modifies aspects of the game that can be concretely
observed by players or admins you should add a changelog. If your change
does NOT meet this description, remove this section. Please note that
maintainers freely reserve the right to remove and add tags should they
deem it appropriate. You can attempt to finagle the system all you want,
but it's best to shoot for clear communication right off the bat. -->

<!-- !! Do not add whitespace in-between the entries, do not change the
tags and do not leave the tags without any entries. -->
:cl:
qol: Subscribers to the server Host, TheDukeofOOK, on Twitch now receive
donator perks!
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [x] You tested this on a local server.
- [x] This code did not runtime during testing.
- [x] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->

---------

Co-authored-by: DESKTOP-M4GD565\LME <polygobttv@gmail.com>
Co-authored-by: CheffieGithub <113442598+CheffieGithub@users.noreply.github.com>