Making all the Jobs follow the guidelines and More (#4497)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request

Title

- Rename Witch Hunter to Monster Hunter and give them a silver amulet if
they are not psydonites,astratans,necrans or pestrans

- Silver amulet is wearable at the wrist

- Give Bloodsucker undead eyes

- Make some migrant jobs being able to contribute to some timelocks, i
forgot about them

- Switch Dwarf Ranger name from Ranger to Dwarf Ranger (i know)

- Noble disguise for the assassin got the foreignnoble lock

- Bloodsucker don't got species limitations besides for the specific
archetypes noble and count

- Assassin and Bloodsucker uses Jobpacks for their archetypes


## Why It's Good For The Game

<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->
It is good

Witch Hunter name didn't fit anymore, silver is a good boon specially if
you are carrying it as a specialist in killing monsters, not wearing one
is dumb

Vampire without night vision? it should be the default but i guess it
can bug

## Changelog

<!-- If your PR modifies aspects of the game that can be concretely
observed by players or admins you should add a changelog. If your change
does NOT meet this description, remove this section. Please note that
maintainers freely reserve the right to remove and add tags should they
deem it appropriate. You can attempt to finagle the system all you want,
but it's best to shoot for clear communication right off the bat. -->

<!-- !! Do not add whitespace in-between the entries, do not change the
tags and do not leave the tags without any entries. -->
:cl:
code: All the jobs follow the code guidelines
bal: Rename Witch Hunter to Monster Hunter and give them a silver amulet
if they are not psydonites,astratans,necrans or pestrans (those already
get pyscrosses)
bal: Silver amulet is wearable at the wrist
bal: GIve bloodsucker undead eyes
change: Dwarf range class is now called Dwarf Ranger instead of Ranger
add:  Make some migrant jobs being able to contribute to some timelocks
bal: Remove the Bloodsucker species limitations and added it
specifically for the archetypes needed
bal: Add species limitation for the Assassin Noble disguise, it is set
the same as for the Noble Pilgrim.
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [x] You tested this on a local server.
- [x] This code did not runtime during testing.
- [x] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->

---------

Co-authored-by: CheffieGithub <113442598+CheffieGithub@users.noreply.github.com>