Adds a towner mason role.  (#4467)
## About The Pull Request
This adds a "Mason" towner role, copy-and-paste of the pilgrim mason
with the exception of the 5% chance to get slightly better skills
(something the towner carpenter has).

## Why It's Good For The Game
Masons are an essential part of any city, and it being pilgrim only is
just plain weird.
Technically this creates an overlap with Artificiers, but honestly I
don't think folks play as artificiers for the mason skills.

## Changelog

:cl:
add: Adds a new mason role for the town. Essentially the pilgrim mason,
except a towner.
/:cl:

## Pre-Merge Checklist
- [x] You tested this on a local server.
- [x] This code did not runtime during testing.
- [x] You documented all of your changes.

---------

Co-authored-by: Somniworld <pablo.capel4@orange.fr>
Co-authored-by: CheffieGithub <113442598+CheffieGithub@users.noreply.github.com>
Co-authored-by: Washington <47699733+wgtjunior743@users.noreply.github.com>